Completed class implementation
So far object shows up properly, just needs more refactoring and functions to use properly
